+++ agent/src/main.rs
@@ -2,17 +2,19 @@ use crate::client::Client;
 
 mod client;
 mod config;
+mod db;
+mod errors;
 mod net;
 
 #[tokio::main]
 async fn main() {
     let ini_path = "config.ini".to_string();
-    let mut client = Client::new(ini_path)
+    Client::new(ini_path)
+        .await
         .map_err(|e| e.to_string())
-        .expect("Failed to start client:");
-    client
+        .expect("Failed to start client")
         .run()
         .await
         .map_err(|e| e.to_string())
-        .expect("Client runtime error:");
+        .expect("Client runtime error");
 }